+++ src/index.js
@@ -4,7 +4,15 @@ import mermaid from "mermaid";
 import katex from "katex";
 import renderMathInElement from "katex/contrib/auto-render";
 import Slugger from "github-slugger";
-import { footnoteList, footnote, renderer, set_repo, slugger, walkTokens } from "./extensions";
+import {
+  footnoteList,
+  footnote,
+  emoji,
+  renderer,
+  set_repo,
+  slugger,
+  walkTokens,
+} from "./extensions";
 
 const search = new URLSearchParams(location.search);
 
@@ -17,7 +25,7 @@ marked.setOptions({
   },
 });
 
-marked.use({ extensions: [footnoteList, footnote], renderer, walkTokens });
+marked.use({ extensions: [footnoteList, footnote, emoji], renderer, walkTokens });
 
 const dark = matchMedia("(prefers-color-scheme: dark)");
 mermaid.initialize({ startOnLoad: false, theme: dark.matches ? "dark" : "default" });